+++ back/src/main/java/itmo/labs/dto/LocationDTO.java
@@ -15,7 +15,7 @@ public class LocationDTO {
 
     private float x;
     @NotNull(message = "Y coordinate cannot be null")
-    private Integer y;
+    private float y;
 
     @NotBlank(message = "Location name cannot be null or empty")
     private String name;
